check fav & comment & follow & report & saved

+++ app/Http/Controllers/Api/FavouriteController.php
@@ -8,11 +8,12 @@ use Validator;
 use App\Models\Favourite;
 use Illuminate\Http\Request;
 use App\Traits\ResponseTrait;
+use App\Traits\MethodconTrait;
 use App\Http\Controllers\Controller;
 
 class FavouriteController extends Controller
 {
-    use ResponseTrait;
+    use ResponseTrait , MethodconTrait;
     //
      /**
      * todo Display a listing of the resource.
@@ -47,13 +48,16 @@ class FavouriteController extends Controller
                 $code = $this->returnCodeAccordingToInput($validator);
                 return $this->returnValidationError($code,$validator);
         }
-        
+
+        $checkfav = $this->checkfav($request->posts_id , auth()->user()->id);
+        if($checkfav !=true){return $this->returnError('Fv001',"Alredy Liked this posts. :( !.."); }
+
         $postfav = Favourite::create([
             'posts_id'  => $request->posts_id,
             'user_id' =>auth()->user()->id,
         ]);
 
-        $msg = " Create successfully .";
+        $msg = " Add Like To This Post successfully .";
         return $this->returnSuccessMessage($msg); 
 
 
@@ -105,13 +109,21 @@ class FavouriteController extends Controller
     public function destroy(Request $request)
     {
         //
+        $msg = " Delete Favourite posts successfully .";
         $postfav = Favourite::where('user_id',auth()->user()->id)
         ->where('posts_id',$request->post_id)->get();
+        
         foreach ($postfav as $fav) {
-          $fav->delete();
-         }
-        $msg = " Delete Favourite posts successfully .";
-        return $this->returnSuccessMessage($msg);  
+            
+            if( isset($fav->id) )
+            {
+              $fav->delete();
+              return $this->returnSuccessMessage($msg);  
+            }
+        }
+        
+        return $this->returnError("FV404" , "Something Wrong :( ! ...");
+
     }
